+++ independent-projects/qute/generator/pom.xml
@@ -46,4 +46,14 @@
         </dependency>
     </dependencies>
 
+    <build>
+        <plugins>
+            <plugin>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <configuration>
+                    <argLine>--add-opens=io.quarkus.qute/io.quarkus.qute=io.quarkus.qute.generator</argLine>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
 </project>
