+++ independent-projects/qute/generator/src/main/java/module-info.java
@@ -1,0 +1,10 @@
+module io.quarkus.qute.generator {
+    requires io.quarkus.gizmo2;
+    requires io.quarkus.qute;
+
+    requires org.jboss.jandex;
+    requires org.jboss.jandex.gizmo2;
+    requires org.jboss.logging;
+
+    exports io.quarkus.qute.generator;
+}
